rename aux.py file to auxiliary.py for windows compatibility. see https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file
